Organize firmware into component folders
